+++ single-exe/packAssets.sh
@@ -4,4 +4,4 @@ sd=$(dirname "$0")
 
 cd "$sd"/..
 
-tar -cvf single-exe/assets.tar runtime README.md CHANGELOG.md node_modules/wasmoon/dist/glue.wasm
+tar -cvf single-exe/assets.tar runtime README.md CHANGELOG.md